consolidate haversine functions
